refactor: move spec to format
