+++ src/lib.rs
@@ -18,9 +18,9 @@ pub use crate::compression::{CompressionMethod, SUPPORTED_COMPRESSION_METHODS};
 pub use crate::datetime::DateTime;
 pub use crate::format::aes::AesMode;
 pub use crate::format::flags::System;
+pub use crate::format::{ZIP64_BYTES_THR, ZIP64_ENTRY_THR};
 pub use crate::read::HasZipMetadata;
 pub use crate::read::{ZipArchive, ZipReadOptions};
-pub use crate::spec::{ZIP64_BYTES_THR, ZIP64_ENTRY_THR};
 pub use crate::write::ZipWriter;
 
 #[cfg(feature = "aes-crypto")]
@@ -38,7 +38,6 @@ mod format;
 mod path;
 pub mod read;
 pub mod result;
-mod spec;
 mod types;
 pub mod write;
 mod zipcrypto;
