+++ src/read/mod.rs
@@ -4,10 +4,10 @@ use crate::compression::CompressionMethod;
 use crate::cp437::FromCp437;
 use crate::datetime::DateTime;
 use crate::extra_fields::{ExtraField, ExtraFields};
+use crate::format::blocks::{CentralDirectoryEndInfo, DataAndPosition, ZipCentralEntryBlock};
 use crate::format::flags::System;
 use crate::format::flags::ZipFlags;
 use crate::result::{ZipError, ZipResult, invalid};
-use crate::spec::{CentralDirectoryEndInfo, DataAndPosition, ZipCentralEntryBlock};
 use crate::types::ZipFileData;
 use indexmap::IndexMap;
 use std::ffi::OsStr;
