+++ src/read/zipfile.rs
@@ -4,14 +4,15 @@ use crate::CompressionMethod;
 use crate::DateTime;
 use crate::HasZipMetadata;
 use crate::ZIP64_BYTES_THR;
+use crate::format::ffi;
 use crate::read::ExtraField;
 use crate::read::RootDirFilter;
 use crate::read::make_writable_dir_all;
 use crate::read::readers::{ZipFileReader, ZipFileSeekReader};
 use crate::result::ZipResult;
 use crate::result::invalid;
+use crate::types::SimpleFileOptions;
 use crate::types::ZipFileData;
-use crate::types::{SimpleFileOptions, ffi};
 use core::mem::replace;
 use std::borrow::Cow;
 use std::ffi::OsStr;
